+++ lib/boruta/adapters/ecto/stores/client_store.ex
@@ -3,33 +3,13 @@ defmodule Boruta.Ecto.ClientStore do
 
   import Boruta.Config, only: [cache_backend: 0]
 
-  alias Boruta.Ecto.Clients
   alias Boruta.Oauth.Client
 
-  @spec get([id: String.t()] | [id: String.t(), secret: String.t()] | [id: String.t(), redirect_uri: String.t()]) ::
+  @spec get_client(String.t()) ::
           {:ok, token :: Boruta.Oauth.Client.t()} | {:error, reason :: String.t()}
-  def get(id: id) do
+  def get_client(id) do
     case get_by_id(id) do
       %Client{} = client -> {:ok, client}
-      nil -> {:error, "Client not cached."}
-      error -> error
-    end
-  end
-  def get(id: id, secret: secret) do
-    with %Client{} = client <- get_by_id(id),
-         :ok <- Clients.check_secret(client, secret) do
-      {:ok, client}
-    else
-      nil -> {:error, "Client not cached."}
-      error -> error
-    end
-  end
-
-  def get(id: id, redirect_uri: redirect_uri) do
-    with %Client{} = client <- get_by_id(id),
-         :ok <- Clients.check_redirect_uri(client, redirect_uri) do
-      {:ok, client}
-    else
       nil -> {:error, "Client not cached."}
       error -> error
     end