+++ test/boruta/oauth/integration/implicit_grant_test.exs
@@ -19,6 +19,7 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
       user = %User{}
       resource_owner = %ResourceOwner{sub: user.id, username: user.email}
       client = insert(:client, redirect_uris: ["https://redirect.uri"])
+      wildcard_redirect_uri_client = insert(:client, redirect_uris: ["https://*.uri"])
       client_without_grant_type = insert(:client, supported_grant_types: [])
 
       client_with_scope =
@@ -30,6 +31,7 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
 
       {:ok,
        client: client,
+       wildcard_redirect_uri_client: wildcard_redirect_uri_client,
        client_with_scope: client_with_scope,
        client_without_grant_type: client_without_grant_type,
        resource_owner: resource_owner}
@@ -154,7 +156,37 @@ defmodule Boruta.OauthTest.ImplicitGrantTest do
 
       redirect_uri = List.first(client.redirect_uris)
 
-      {:authorize_success,
+      assert {:authorize_success,
+       %AuthorizeResponse{
+         type: type,
+         access_token: value,
+         expires_in: expires_in,
+         redirect_uri: ^redirect_uri
+       }} =
+        Oauth.authorize(
+          %Plug.Conn{
+            query_params: %{
+              "response_type" => "token",
+              "client_id" => client.id,
+              "redirect_uri" => redirect_uri
+            }
+          },
+          resource_owner,
+          ApplicationMock
+        )
+
+      assert type == :token
+      assert value
+      assert expires_in
+    end
+
+    test "returns a token with a wildcard client redirect uri", %{wildcard_redirect_uri_client: client, resource_owner: resource_owner} do
+      ResourceOwners
+      |> expect(:get_by, fn _params -> {:ok, resource_owner} end)
+
+      redirect_uri = "https://wildcard-redirect-uri.uri"
+
+      assert {:authorize_success,
        %AuthorizeResponse{
          type: type,
          access_token: value,