+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - better Ecto errors management
     - remove padding from pkce code challenge checks
     - reduce resource_owners adapter calls
+  ### Added
+    - Domain wildcard for client redirect_uris
 ## [2.0.0-rc.0]
   ### Added
     - [OpenID Connect core 1.0](https://openid.net/specs/openid-connect-core-1_0.html) integration